1,修改玩法的中奖匹配

+++ src/main/java/com/jll/game/playtypefacade/EleIn5QszuxZlPlayTypeFacadeImpl.java
@@ -152,8 +152,8 @@ public class EleIn5QszuxZlPlayTypeFacadeImpl extends DefaultPlayTypeFacadeImpl  
 		
 		
 		betAmount = MathUtil.multiply(winningBetAmount, times, Float.class);
-		betAmount = MathUtil.multiply(betAmount, monUnit, Float.class);
-		maxWinAmount = MathUtil.multiply(betAmount, singleBettingPrize, Float.class);
+		betAmount = MathUtil.multiply(betAmount, monUnit.floatValue(), Float.class);
+		maxWinAmount = MathUtil.multiply(betAmount, singleBettingPrize.floatValue(), Float.class);
 		
 		return new BigDecimal(maxWinAmount);
 	}
